+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\index.js":"1","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\App.js":"2","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\BookForm.js":"3","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\BookList.js":"4","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\Book.js":"5","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\reducers\\index.js":"6","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\reducers\\book.js":"7"},{"size":234,"mtime":1611957479584,"results":"8","hashOfConfig":"9"},{"size":464,"mtime":1612279783352,"results":"10","hashOfConfig":"9"},{"size":438,"mtime":1612278813505,"results":"11","hashOfConfig":"9"},{"size":308,"mtime":1612279997076,"results":"12","hashOfConfig":"9"},{"size":348,"mtime":1612278813505,"results":"13","hashOfConfig":"9"},{"size":168,"mtime":1612278813504,"results":"14","hashOfConfig":"9"},{"size":472,"mtime":1612281734387,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"1u47te5",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\index.js",[],["31","32","33","34","35","36"],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\App.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\BookForm.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\BookList.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\Book.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\reducers\\index.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\reducers\\book.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},{"ruleId":"45","replacedBy":"46"},{"ruleId":"47","replacedBy":"48"},"lines-around-directive",["49"],"no-spaced-func",["50"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements","func-call-spacing"]
+[{"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/index.js":"1","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/App.js":"2","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/BookForm.js":"3","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/BookList.js":"4","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/reducers/index.js":"5","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/Book.js":"6","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/reducers/book.js":"7"},{"size":763,"mtime":1612285632470,"results":"8","hashOfConfig":"9"},{"size":254,"mtime":1612282382714,"results":"10","hashOfConfig":"9"},{"size":438,"mtime":1612217260142,"results":"11","hashOfConfig":"9"},{"size":711,"mtime":1612286974299,"results":"12","hashOfConfig":"9"},{"size":168,"mtime":1612210881932,"results":"13","hashOfConfig":"9"},{"size":348,"mtime":1612286824072,"results":"14","hashOfConfig":"9"},{"size":254,"mtime":1612285070386,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"5o415q",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/index.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/App.js",[],["31","32","33","34","35","36"],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/BookForm.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/BookList.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/reducers/index.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/Book.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/reducers/book.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},{"ruleId":"45","replacedBy":"46"},{"ruleId":"47","replacedBy":"48"},"lines-around-directive",["49"],"no-spaced-func",["50"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements","func-call-spacing"]
