+++ src/reducers/book.js
@@ -1,19 +1,4 @@
-const initialState = {
-  books: [
-    {
-      id: 1,
-      title: 'Lord of the rings',
-      category: 'Action',
-    },
-    {
-      id: 2,
-      title: 'Harry Potter',
-      category: 'Kids',
-    },
-  ],
-};
-
-const BookReducer = (state = initialState, action) => {
+const BookReducer = (state = [], action) => {
   switch (action.type) {
     case 'CREATE_BOOK': return {
       ...state,
